refactor: 마커 컴포넌트 폴더 구조 개선
[#914]
