+++ frontend/src/components/google-maps/map/CarFfeineMap.tsx
@@ -5,7 +5,9 @@ import DataDownloader from '@ui/DataDownloader';
 import CarFfeineMapListener from './CarFfeineListener';
 
 const UserFilterListener = lazy(() => import('./UserFilterListener'));
-const MarkersContainers = lazy(() => import('@marker/MarkerContainers'));
+const MarkersContainers = lazy(
+  () => import('@marker/components/MarkerContainers/MarkerContainers')
+);
 const ToastContainer = lazy(() => import('@ui/ToastContainer'));
 const ClientStationFilters = lazy(() => import('@ui/ClientStationFilters'));
 const MapController = lazy(() => import('@ui/MapController'));
