+++ frontend/src/components/google-maps/marker/components/LargeDeltaAreaMarkerContainer/LargeDeltaAreaMarkerContainer.tsx
