+++ frontend/src/components/google-maps/marker/components/LargeDeltaAreaMarkerContainer/components/StyledClusterMarker.tsx
