+++ frontend/src/components/google-maps/marker/components/LargeDeltaAreaMarkerContainer/index.ts
