+++ frontend/src/components/google-maps/marker/components/MaxDeltaAreaMarkerContainer/components/RegionMarker.stories.tsx
