+++ frontend/src/components/google-maps/marker/components/SmallMediumDeltaAreaMarkerContainer/components/CarFfeineMarker/CarFfeine.stories.tsx
