+++ frontend/src/components/google-maps/marker/components/SmallMediumDeltaAreaMarkerContainer/components/marker/DefaultMarkerRenderer.tsx
