+++ frontend/src/components/google-maps/marker/components/SmallMediumDeltaAreaMarkerContainer/constants/index.ts
