+++ frontend/src/components/google-maps/marker/components/SmallMediumDeltaAreaMarkerContainer/index.ts
