+++ frontend/src/components/ui/MapController.tsx
@@ -3,8 +3,8 @@ import { css } from 'styled-components';
 
 import { BiCurrentLocation } from 'react-icons/bi';
 
-import { useClusterMarkers } from '@marker/LargeDeltaAreaMarkerContainer/hooks/useClusterMarkers';
-import { useStationMarkers } from '@marker/SmallMediumDeltaAreaMarkerContainer/hooks/useStationMarkers';
+import { useClusterMarkers } from '@marker/components/LargeDeltaAreaMarkerContainer/hooks/useClusterMarkers';
+import { useStationMarkers } from '@marker/components/SmallMediumDeltaAreaMarkerContainer/hooks/useStationMarkers';
 
 import { googleMapActions } from '@stores/google-maps/googleMapStore';
